pyproject.toml
[project]
name = "core"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "python-jose>=3.5.0",
    "typing>=3.10.0.0",
    "uvicorn>=0.35.0",
]


[tool.ruff]
line-length = 120
target-version = "py310"
lint.select = ["E","F","I","UP"]
lint.ignore = ["E501"]
format.quote-style = "double"
format.indent-style = "space"
format.skip-magic-trailing-comma = true
